Add a basic GUI to Wessy Use JavaFX technology to implement the GUI.
