+++ src/main/java/wessy/Wessy.java
@@ -3,6 +3,7 @@ package wessy;
 import java.io.IOException;
 import java.time.format.DateTimeParseException;
 
+import wessy.components.*;
 import wessy.task.Task;
 
 import wessy.exceptions.WessyException;
@@ -39,6 +40,10 @@ public class Wessy {
             loadedTasks = new TaskList();
         }
         tasks = loadedTasks;
+    }
+
+    public Wessy() {
+        this("data/savedTasks.txt");
     }
 
     /**
@@ -122,7 +127,6 @@ public class Wessy {
             } catch (IOException ioe) {
                 ui.handleException("There is some issue in the input-output operation.");
                 ioe.printStackTrace();
-
             } catch (WessyException we) {
                 ui.handleException(we.toString());
             } catch (Exception ex) {
@@ -141,10 +145,10 @@ public class Wessy {
     }
 
     /**
-     * A helper function
+     *
      */
-    private void startsUp() {
-        ui.printWelcomeMessage(tasks.printAsStr(), tasks.getSize());
+    public String startsUp() {
+        return ui.printWelcomeMessage(tasks.printAsStr());
     }
 
     /**
@@ -172,4 +176,82 @@ public class Wessy {
         storage.save(tasks.saveAsStr());
     }
 
+
+
+    public String respond(String userInput) {
+        try {
+            CmdType cmd = Parser.getCmd(userInput);
+            if (cmd == null) {
+                throw new CommandNotFoundException();
+            }
+            UserInputChecker.checkSpacingAftCmd(userInput, cmd);
+
+            switch (cmd) {
+
+                case BYE:
+                    return ui.printByeMessage();
+                case LIST:
+                    return ui.printListOrFindMessage(tasks.printAsStr(), true);
+                case TODO:
+                    // Fallthrough
+                case DEADLINE:
+                    // Fallthrough
+                case EVENT:
+
+                    UserInputChecker.checkMissingInput(userInput, cmd);
+                    UserInputChecker.checkMissingKeyword(userInput, cmd);
+                    if (cmd == CmdType.DEADLINE) {
+                        UserInputChecker.checkDeadlineMissingInput(userInput);
+                    } else if (cmd == CmdType.EVENT) {
+                        UserInputChecker.checkEventMissingInput(userInput);
+                    }
+
+                    String[] taskComponents = Parser.getTaskComponents(userInput, cmd);
+                    Task newTask = tasks.add(taskComponents);
+                    saveToStorage();
+                    return ui.printAddedMessage(newTask, tasks.getSize());
+
+                case MARK:
+                case UNMARK:
+                    checkBeforeParse(userInput, cmd);
+
+                    boolean isMark = cmd == CmdType.MARK;
+                    Task updatedTask = tasks.markOrUnmark(Parser.parseInt(userInput, cmd), isMark);
+                    saveToStorage();
+                    return ui.printMarkUnmarkMessage(updatedTask, isMark);
+
+                case DELETE:
+                    checkBeforeParse(userInput, cmd);
+
+                    Task deletedTask = tasks.delete(Parser.parseInt(userInput, cmd));
+                    saveToStorage();
+                    return ui.printDeleteMessage(deletedTask, tasks.getSize());
+
+                case FIND:
+                    String target = userInput.substring(cmd.getStrLength() + 1);
+                    return ui.printListOrFindMessage(tasks.find(target), false);
+
+                case CLEAR:
+                    tasks.clear();
+                    saveToStorage();
+                    return ui.printClearMessage();
+                    // Fallthrough
+            }
+
+        } catch (DateTimeParseException dtpe) {
+            return ui.handleException("Please enter the date (and time, if any) in the correct format.");
+        } catch (SecurityException se) {
+            se.printStackTrace();
+            return ui.handleException("You do not have the permission to access the file.");
+        } catch (IOException ioe) {
+            ioe.printStackTrace();
+            return ui.handleException("There is some issue in the input-output operation.");
+
+        } catch (WessyException we) {
+            return ui.printMessage(we.toString());
+        } catch (Exception ex) {
+            return ui.printMessage(ex.getMessage());
+        }
+        return ui.printMessage(new CommandNotFoundException().toString());
+    }
 }
