+++ src/main/resources/view/DialogBox.fxml
@@ -1,0 +1,17 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<?import javafx.scene.control.Label?>
+<?import javafx.scene.image.ImageView?>
+<?import javafx.scene.layout.HBox?>
+<?import javafx.scene.layout.Pane?>
+
+<fx:root alignment="TOP_RIGHT" maxHeight="1.7976931348623157E308" maxWidth="1.7976931348623157E308" prefWidth="400.0" minHeight="150"
+         type="javafx.scene.layout.HBox" xmlns="http://javafx.com/javafx/8.0.171" xmlns:fx="http://javafx.com/fxml/1"
+         style="-fx-background-color:rgb(182, 252, 213);-fx-padding: 10 10 10 10">
+    <children>
+        <Label fx:id="dialog" text="Label" wrapText="true" maxWidth="220" alignment="CENTER_LEFT" style="-fx-padding: 10 10 10 10;"/>
+        <Pane fx:id="displayPicture" style="-fx-padding: 5 5 5 5;">
+            <ImageView fx:id="iv" fitHeight="99.0" fitWidth="99.0" pickOnBounds="true" preserveRatio="true"/>
+        </Pane>
+    </children>
+</fx:root>
